update session function

+++ app/src/main/java/com/example/owlfriend/domain/Session.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class Session(
     val relatedToSubject: String,
     val date: Long,
-    val duration: Float,
+    val duration: Long,
     val sessionSubjectId: Int,
     @PrimaryKey(autoGenerate = true)
     val sessionId: Int? = null
